@@ -121,28 +121,30 @@ func setupLocalDefaults(thread *starlark.Thread) error {
 func newPredeclareds(restrictedMode []bool) starlark.StringDict {
 
 	dict := starlark.StringDict{
-		identifiers.os:                setupOSStruct(),
-		identifiers.crashdCfg:         starlark.NewBuiltin(identifiers.crashdCfg, crashdConfigFn),
-		identifiers.sshCfg:            starlark.NewBuiltin(identifiers.sshCfg, SshConfigFn),
-		identifiers.hostListProvider:  starlark.NewBuiltin(identifiers.hostListProvider, hostListProvider),
-		identifiers.resources:         starlark.NewBuiltin(identifiers.resources, resourcesFunc),
-		identifiers.archive:           starlark.NewBuiltin(identifiers.archive, archiveFunc),
-		identifiers.run:               starlark.NewBuiltin(identifiers.run, runFunc),
-		identifiers.runLocal:          starlark.NewBuiltin(identifiers.runLocal, runLocalFunc),
-		identifiers.progAvailLocal:    starlark.NewBuiltin(identifiers.progAvailLocal, progAvailLocalFunc),
-		identifiers.capture:           starlark.NewBuiltin(identifiers.capture, captureFunc),
-		identifiers.captureLocal:      starlark.NewBuiltin(identifiers.capture, captureLocalFunc),
-		identifiers.copyFrom:          starlark.NewBuiltin(identifiers.copyFrom, copyFromFunc),
-		identifiers.copyTo:            starlark.NewBuiltin(identifiers.copyTo, copyToFunc),
-		identifiers.kubeCfg:           starlark.NewBuiltin(identifiers.kubeCfg, KubeConfigFn),
-		identifiers.kubeCapture:       starlark.NewBuiltin(identifiers.kubeGet, KubeCaptureFn),
-		identifiers.kubeGet:           starlark.NewBuiltin(identifiers.kubeGet, KubeGetFn),
-		identifiers.kubeExec:          starlark.NewBuiltin(identifiers.kubeExec, KubeExecFn),
-		identifiers.kubeNodesProvider: starlark.NewBuiltin(identifiers.kubeNodesProvider, KubeNodesProviderFn),
-		identifiers.capvProvider:      starlark.NewBuiltin(identifiers.capvProvider, CapvProviderFn),
-		identifiers.capaProvider:      starlark.NewBuiltin(identifiers.capaProvider, CapaProviderFn),
-		identifiers.setDefaults:       starlark.NewBuiltin(identifiers.setDefaults, SetDefaultsFunc),
-		identifiers.log:               starlark.NewBuiltin(identifiers.log, logFunc),
+		identifiers.os:                    setupOSStruct(),
+		identifiers.crashdCfg:             starlark.NewBuiltin(identifiers.crashdCfg, crashdConfigFn),
+		identifiers.sshCfg:                starlark.NewBuiltin(identifiers.sshCfg, SshConfigFn),
+		identifiers.hostListProvider:      starlark.NewBuiltin(identifiers.hostListProvider, hostListProvider),
+		identifiers.resources:             starlark.NewBuiltin(identifiers.resources, resourcesFunc),
+		identifiers.archive:               starlark.NewBuiltin(identifiers.archive, archiveFunc),
+		identifiers.run:                   starlark.NewBuiltin(identifiers.run, runFunc),
+		identifiers.runLocal:              starlark.NewBuiltin(identifiers.runLocal, runLocalFunc),
+		identifiers.progAvailLocal:        starlark.NewBuiltin(identifiers.progAvailLocal, progAvailLocalFunc),
+		identifiers.capture:               starlark.NewBuiltin(identifiers.capture, captureFunc),
+		identifiers.captureLocal:          starlark.NewBuiltin(identifiers.capture, captureLocalFunc),
+		identifiers.copyFrom:              starlark.NewBuiltin(identifiers.copyFrom, copyFromFunc),
+		identifiers.copyTo:                starlark.NewBuiltin(identifiers.copyTo, copyToFunc),
+		identifiers.kubeCfg:               starlark.NewBuiltin(identifiers.kubeCfg, KubeConfigFn),
+		identifiers.kubeCapture:           starlark.NewBuiltin(identifiers.kubeGet, KubeCaptureFn),
+		identifiers.kubeGet:               starlark.NewBuiltin(identifiers.kubeGet, KubeGetFn),
+		identifiers.kubeExec:              starlark.NewBuiltin(identifiers.kubeExec, KubeExecFn),
+		identifiers.kubeNodesProvider:     starlark.NewBuiltin(identifiers.kubeNodesProvider, KubeNodesProviderFn),
+		identifiers.capvProvider:          starlark.NewBuiltin(identifiers.capvProvider, CapvProviderFn),
+		identifiers.capaProvider:          starlark.NewBuiltin(identifiers.capaProvider, CapaProviderFn),
+		identifiers.kcpProvider:           starlark.NewBuiltin(identifiers.kcpProvider, KcpProviderFn),
+		identifiers.setDefaults:           starlark.NewBuiltin(identifiers.setDefaults, SetDefaultsFunc),
+		identifiers.kubePortForwardConfig: starlark.NewBuiltin(identifiers.kubePortForwardConfig, KubePortForwardrFn),
+		identifiers.log:                   starlark.NewBuiltin(identifiers.log, logFunc),
 	}
 
 	if len(restrictedMode) > 0 && restrictedMode[0] {
